+++ src/exports.jl
@@ -99,6 +99,7 @@ export
 
 # Dofs
     DofHandler,
+    MixedDofHandler, # only for getting an error message redirecting to DofHandler
     close!,
     ndofs,
     ndofs_per_cell,
@@ -109,7 +110,6 @@ export
     dof_range,
     renumber!,
     DofOrder,
-    MixedDofHandler,
     FieldHandler,
     Field,
     reshape_to_nodes,